Use unique release name for test bundles
Currently when running the autoupdate and autodeploy jobs, they all have the same branch name which results in the same release name. When multiple jobs want to create a release with the same name, it causes a conflict and it fails.

As a solution, we are generating a unique release name for creating bundles.